requirements.txt
appdirs==1.4.4
astroid==2.5.1
autopep8==1.5.6
bidict==0.21.2
blinker==1.4
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.4.2
pylint==2.7.2
pymongo==3.11.3
python-dotenv==0.16.0
python-engineio==3.13.2
python-snappy==0.6.0
python-socketio==4.6.0
regex==2021.3.17
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
winkerberos==0.7.0
wrapt==1.12.1
